继续添加测试函数。

+++ StockAnalysis/source/AmericaMarket/WorldStock.h
@@ -30,7 +30,7 @@ public:
   void UpdateEPSSurprise(vector<CEPSSurprisePtr>& vEPSSurprise);
 
   void UpdateDayLineStartEndDate(void);
-  long GetDayLineSize(void) noexcept { return m_vDayLine.size(); }
+  long GetDayLineSize(void) noexcept { return (long)m_vDayLine.size(); }
   CDayLinePtr GetDayLine(long lIndex) { return m_vDayLine.at(lIndex); }
   void UnloadDayLine(void) { m_vDayLine.resize(0); }
